# apps/server-go/api/openapi.yaml
openapi: 3.1.0
info:
  title: FixPro API
  version: 1.0.0-go
servers:
  - url: http://localhost:8080
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: {type: string, pattern: '^[1-9][0-9]*$'}
  securitySchemes:
    adminSession: {type: apiKey, in: cookie, name: fixpro_admin_session}
    adminBasic: {type: http, scheme: basic, description: Deprecated local compatibility only}
...
paths:
...
  /api/v1/mini/orders:
    post:
      security: [{miniBearer: []}]
      parameters: [{name: Idempotency-Key, in: header, required: true, schema: {type: string, maxLength: 128}}]
      responses: {'200': {description: Order created or replayed}}
  /api/v1/mini/orders/{id}:
    get: {security: [{miniBearer: []}], parameters: [{$ref: '#/components/parameters/Id'}], responses: {'200': {description: Customer order detail and work order progress}}}
...